# grapevine-client/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "grapevine API",
    "description": "Decentralized Data Feeds Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/v1/transactions/{id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get transaction by ID",
        "description": "Retrieve a specific transaction by its UUID. Public endpoint - no authentication required.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Transaction UUID"
            },
            "required": true,
            "description": "Transaction UUID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "piid": {
                      "type": "string",
                      "nullable": true,
                      "format": "uuid"
                    },
                    "payer": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    "pay_to": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    "amount": {
                      "type": "string",
                      "maxLength": 20
                    },
                    "asset": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "entry_id": {
                      "type": "string",
                      "nullable": true,
                      "format": "uuid"
                    },
                    "transaction_hash": {
                      "type": "string",
                      "minLength": 66,
                      "maxLength": 66,
                      "pattern": "^0x[0-9a-fA-F]{64}$"
                    },
                    "created_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    }
                  },
                  "required": [
                    "id",
                    "piid",
                    "payer",
                    "pay_to",
                    "amount",
                    "asset",
                    "entry_id",
                    "transaction_hash",
                    "created_at"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/hash/{hash}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get transaction by hash",
        "description": "Retrieve a transaction by its blockchain transaction hash (0x-prefixed, 66 characters). Public endpoint - no authentication required.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 66,
              "maxLength": 66,
              "description": "Blockchain transaction hash (66 characters with 0x prefix)",
              "example": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
            },
            "required": true,
            "description": "Blockchain transaction hash (66 characters with 0x prefix)",
            "name": "hash",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "piid": {
                      "type": "string",
                      "nullable": true,
                      "format": "uuid"
                    },
                    "payer": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    "pay_to": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    },
                    "amount": {
                      "type": "string",
                      "maxLength": 20
                    },
                    "asset": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "entry_id": {
                      "type": "string",
                      "nullable": true,
                      "format": "uuid"
                    },
                    "transaction_hash": {
                      "type": "string",
                      "minLength": 66,
                      "maxLength": 66,
                      "pattern": "^0x[0-9a-fA-F]{64}$"
                    },
                    "created_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    }
                  },
                  "required": [
                    "id",
                    "piid",
                    "payer",
                    "pay_to",
                    "amount",
                    "asset",
                    "entry_id",
                    "transaction_hash",
                    "created_at"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}